+++ components/src/components/ins-carousel/CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## Changes from V1
 - Methods are now asynchronous
-- Update `toggle` to `insToggle` to avoid DOM event conflict
+- Update `slide` to `insSlide` to avoid DOM event conflict