+++ components/src/components/ins-checkbox-card/CHANGELOG.md
@@ -1,6 +1,6 @@
 # v2.0.0
 
 ## Changes from V1
-- Updated event `onClickInsCheckboxCard` to `clickInsCheckboxCard` to describe the event better
-- Updated event `valueChange` to `changeValue` to avoid conflict with the native DOM event name
+- Updated event `onClickInsCheckboxCard` to `insClick` to describe the event better
+- Updated event `valueChange` to `insValueChange` to avoid conflict with the native DOM event name
 - Added a default value of `#1e86e3` for `selectedColor` property