+++ components/src/components/ins-credit-card/CHANGELOG.md
@@ -1,2 +1,6 @@
 # v2.0.0
-Start of documentation
+
+## Changes from V1
+- Update `insCreditCardClicked` to `insClick` to avoid DOM event conflict
+- Update `insCreditCardClosed` to `insClose` to avoid DOM event conflict
+- Include SVG logos