+++ components/src/components/ins-date-time/CHANGELOG.md
@@ -1,5 +1,6 @@
 # v2.0.0
 
 ## Changes from V1
-- Updated event `onpick` to `pick` to describe the event better
-- Updated event `valueChange` to `changeValue` to avoid conflict with the native DOM event name
+- Updated event `onpick` to `insInput` to avoid conflict with the native DOM event name
+- Updated event `valueChange` to `insValueChange` to avoid conflict with the native DOM event name
+- Added support for date and time only with `mode` attribute