+++ components/src/components/ins-datepicker/CHANGELOG.md
@@ -1,5 +1,6 @@
 # v2.0.0
 
 ## Changes from V1
-- Updated event `onDateChange` to `pick` to be consistent with `ins-date-time`
-- Updated event `valueChange` to `changeValue` to avoid conflict with the native DOM event name
+- Updated event `onpick` to `insInput` to avoid conflict with the native DOM event name
+- Updated event `valueChange` to `insValueChange` to avoid conflict with the native DOM event name
+- Use ins-date-time component as source to utilise plugin