+++ components/src/components/ins-select-option/CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## Changes from V1
 - Methods are now asynchronous
-- Update `toggle` to `insToggle` to avoid DOM event conflict
+- Removed unused event